refactor(work): split work-state.js into work-state/ submodules + de-allowlist
Quality-allowlist burndown for work-state.js. The file was an allowlisted
1213-line module violating max-lines (file), max-lines-per-function (main 156,
autoCompleteCheckpointTasks 124), cyclomatic (main 44, autoCompleteCheckpointTasks
23, autoInitTdd 14, readTaskInitDescriptors 11) and cognitive limits.

Behavior-preserving split: the implementation moves into cohesive work-state/
submodules; work-state.js stays the single public entry point (facade) — it
installs the CLI process guards, wires the submodules + _setParent injection, and
re-exports the same 30-symbol surface every existing consumer imports.

  - work-state/core.js        config bootstrap, IO primitives, autoInitTdd
                              (path resolution split out -> cyclo ok)
  - work-state/checkpoints.js autoCompleteCheckpointTasks decomposed into
                              report-read / tasks.md re-check / verdict / linkage
                              helpers; security regexes + fail-closed logic verbatim
  - work-state/steps.js       step/check/error/complete/resume/format helpers
                              (statusIcon + render helpers -> cognitive ok)
  - work-state/subtasks.js    subtask state functions (verbatim)
  - work-state/tasks.js       task-progress + fix-rounds (shared loadTasksMeta /
                              resolveCurrentTask helpers remove the duplicate block)
  - work-state/task-init.js   stdin hardening + descriptor parsing
                              (parseDescriptorPayload split -> cyclo ok)
  - work-state/cli.js         main() -> thin dispatch over a HANDLERS table
                              (cyclo 44->ok, 156 lines->ok); per-command
                              stdout/stderr/exit codes byte-for-byte unchanged

work-state.js 1213 -> 167 lines; every function now <=80 lines, cyclomatic <=10,
cognitive <=15. Removes work-state.js from .quality-exceptions. All 769 work/
tests pass; full static-quality gate is green.
