+++ .quality-exceptions
@@ -42,7 +42,6 @@
 
 plugins/work/scripts/workflows/work/scripts/follow-up-pr.js
 plugins/work/scripts/workflows/lib/hooks/enforce-step-workflow.js
-plugins/work/scripts/workflows/work/work-state.js
 plugins/work/scripts/workflows/work/workflow-definition.js
 plugins/work/scripts/workflows/work-implement/task-next.js
 plugins/work/scripts/workflows/work/work-next.js
